Time Dimensions: engine + tab UI (Feature 4.3)
Engine: 8 TD tiers (amount/bought/cost) bought with EP; the original cost curve
(per-tier base/mult, threshold bumps x1.5/x2.2 at 1.8e308 / 1e1300, e6000
super-scaling); production chain TD8->..->TD1 -> Time Shards; FreeTickspeed
.fromShards port (1.33 mult, 300k softcap, Newton inversion past it) feeding
total_tick_gained. Tickspeed moves to Decimal (bought + free upgrades would
underflow f64). Eternity reset zeroes shards/free upgrades and returns TD
amounts to the bought base. TD5-8 modelled but locked: their unlocks are
Dilation studies (Phase 5) - decomposition doc corrected.

Save: dimensions.time[] (cost derived on load), timeShards, totalTickGained.

GUI: Time Dimensions subtab under Eternity (vendored Modern rows, shard/free-
tickspeed readouts, Max all), buy/buy-max/max-all commands, first-eternity
navigation to the tab.

+++ crates/ad-core/src/observed.rs
@@ -55,7 +55,7 @@ pub struct ObservedTickspeedState {
     /// Cost multiplier per purchase.
     pub cost_multiplier: Decimal,
     /// Current tickspeed interval in milliseconds.
-    pub tickspeed_ms: f64,
+    pub tickspeed_ms: Decimal,
     /// Production multiplier from tickspeed.
     pub tickspeed_effect: Decimal,
 }
